requirement.txt
aiofiles==23.2.1
aiohappyeyeballs==2.4.0
aiohttp==3.10.6
aiosignal==1.3.1
alabaster==1.0.0
annotated-types==0.7.0
anthropic==0.34.2
anyio==4.6.0
arrow==1.3.0
asgiref==3.8.1
asttokens==2.4.1
async-timeout==4.0.3
attrs==24.2.0
azure-ai-documentintelligence==1.0.0b4
azure-core==1.31.0
babel==2.16.0
backoff==2.2.1
bcrypt==4.2.0
beautifulsoup4==4.12.3
binaryornot==0.4.4
black==24.8.0
boto3==1.35.26
botocore==1.35.26
build==1.2.2
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
cfgv==3.4.0
chardet==5.2.0
charset-normalizer==3.3.2
chroma-hnswlib==0.7.6
chromadb==0.5.7
click==8.1.7
cohere==5.9.4
coloredlogs==15.0.1
colorlog==6.8.2
contourpy==1.3.0
cookiecutter==2.6.0
coverage==7.6.1
cryptography==43.0.1
cycler==0.12.1
dataclasses-json==0.6.7
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
deprecation==2.1.0
dirtyjson==1.0.8
diskcache==5.6.3
distlib==0.3.8
distro==1.9.0
docutils==0.21.2
duckduckgo_search==6.1.12
durationpy==0.7
elastic-transport==8.15.0
elasticsearch==8.13.2
environs==9.5.0
et-xmlfile==1.1.0
exceptiongroup==1.2.2
executing==2.1.0
fast-langdetect==0.2.1
fastapi==0.112.1
fastavro==1.9.7
fastembed==0.3.6
fasttext-wheel==0.9.2
ffmpy==0.4.0
filelock==3.16.1
flake8==7.1.1
flatbuffers==24.3.25
fonttools==4.54.1
frozenlist==1.4.1
fsspec==2024.9.0
google-auth==2.35.0
googleapis-common-protos==1.65.0
googlesearch-python==1.2.5
gradio==4.39.0
gradio_client==1.1.1
greenlet==3.1.1
grpcio==1.66.1
grpcio-tools==1.62.3
h11==0.14.0
h2==4.1.0
hpack==4.0.0
html2text==2024.2.26
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.25.1
humanfriendly==10.0
hyperframe==6.0.1
identify==2.6.1
idna==3.10
imagesize==1.4.1
importlib_metadata==8.4.0
importlib_resources==6.4.5
iniconfig==2.0.0
ipython==8.27.0
isodate==0.6.1
jedi==0.19.1
Jinja2==3.1.4
jiter==0.5.0
jmespath==1.0.1
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.4.7
-e git+https://github.com/Cinnamon/kotaemon@79b309396bfdad33fb44dd40324667f2e4d83a26#egg=kotaemon&subdirectory=libs/kotaemon
-e git+https://github.com/Cinnamon/kotaemon@79b309396bfdad33fb44dd40324667f2e4d83a26#egg=ktem&subdirectory=libs/ktem
kubernetes==31.0.0
lancedb==0.13.0
langchain==0.2.15
langchain-anthropic==0.1.23
langchain-cohere==0.2.4
langchain-community==0.2.11
langchain-core==0.2.41
langchain-experimental==0.0.64
langchain-openai==0.1.25
langchain-text-splitters==0.2.4
langsmith==0.1.128
linkify-it-py==2.0.3
llama-cloud==0.1.0
llama-hub==0.0.79.post1
llama-index==0.10.68
llama-index-agent-openai==0.2.9
llama-index-cli==0.1.13
llama-index-core==0.10.68.post1
llama-index-embeddings-openai==0.1.11
llama-index-indices-managed-llama-cloud==0.2.7
llama-index-legacy==0.9.48.post3
llama-index-llms-openai==0.1.31
llama-index-multi-modal-llms-openai==0.1.9
llama-index-program-openai==0.1.7
llama-index-question-gen-openai==0.1.3
llama-index-readers-file==0.1.33
llama-index-readers-llama-parse==0.1.6
llama-index-vector-stores-chroma==0.1.10
llama-index-vector-stores-lancedb==0.1.7
llama-index-vector-stores-milvus==0.1.23
llama-index-vector-stores-qdrant==0.2.17
llama-parse==0.4.9
llama_cpp_python==0.2.7
loguru==0.7.2
lxml==5.3.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.22.0
matplotlib==3.9.2
matplotlib-inline==0.1.7
mccabe==0.7.0
mdit-py-plugins==0.4.2
mdurl==0.1.2
milvus-lite==2.4.10
mmh3==4.1.0
monotonic==1.6
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.3
nltk==3.9.1
nodeenv==1.9.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.68
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
onnx==1.16.2
onnxruntime==1.19.2
openai==1.47.1
openpyxl==3.1.5
opentelemetry-api==1.27.0
opentelemetry-exporter-otlp-proto-common==1.27.0
opentelemetry-exporter-otlp-proto-grpc==1.27.0
opentelemetry-instrumentation==0.48b0
opentelemetry-instrumentation-asgi==0.48b0
opentelemetry-instrumentation-fastapi==0.48b0
opentelemetry-proto==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
opentelemetry-util-http==0.48b0
orjson==3.10.7
overrides==7.7.0
packaging==24.1
pandas==2.2.3
parameterized==0.9.0
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.3.6
plotly==5.24.1
pluggy==1.5.0
portalocker==2.10.1
posthog==3.6.6
pre-commit==3.8.0
prompt_toolkit==3.0.47
protobuf==4.25.5
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py==1.11.0
pyaml==23.12.0
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pybind11==2.13.6
pycodestyle==2.12.1
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pydub==0.25.1
pyflakes==3.2.0
Pygments==2.18.0
pylance==0.17.0
pymilvus==2.4.6
PyMuPDF==1.24.10
PyMuPDFb==1.24.10
pyparsing==3.1.4
pypdf==4.2.0
PyPika==0.48.9
pyproject_hooks==1.1.0
pyreqwest_impersonate==0.5.3
PyStemmer==2.2.0.1
pytest==8.3.3
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-docx==1.1.2
python-dotenv==1.0.1
python-multipart==0.0.10
python-slugify==8.0.4
pytz==2024.2
PyYAML==6.0.2
qdrant-client==1.11.3
regex==2024.9.11
requests==2.32.3
requests-oauthlib==2.0.0
retry==0.9.2
retrying==1.3.4
rich==13.8.1
robust-downloader==0.0.2
rsa==4.9
ruff==0.6.7
s3transfer==0.10.2
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
semantic-version==2.10.0
sentence-transformers==3.1.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
snowballstemmer==2.2.0
soupsieve==2.6
Sphinx==8.0.2
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
SQLAlchemy==2.0.35
sqlmodel==0.0.22
stack-data==0.6.3
starlette==0.38.6
striprtf==0.0.26
sympy==1.13.3
tabulate==0.9.0
tantivy==0.22.0
tenacity==8.2.3
text-unidecode==1.3
textual==0.80.1
theflow==0.8.6
threadpoolctl==3.5.0
tiktoken==0.7.0
tokenizers==0.19.1
tomli==2.0.1
tomlkit==0.12.0
torch==2.4.1
tqdm==4.66.5
traitlets==5.14.3
transformers==4.44.2
triton==3.0.0
trogon==0.5.0
typer==0.12.5
types-python-dateutil==2.9.0.20240906
types-requests==2.32.0.20240914
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
uc-micro-py==1.0.3
ujson==5.10.0
urllib3==2.2.3
uvicorn==0.30.6
uvloop==0.20.0
virtualenv==20.26.5
watchfiles==0.24.0
wcwidth==0.2.13
websocket-client==1.8.0
websockets==11.0.3
wikipedia==1.4.0
wrapt==1.16.0
yarl==1.12.1
zipp==3.20.2
